Migrate TArraySpec to ZIO test.
